External false discovery rate ('FDR') control methods

Procedures used by focr to control the FDR in the second stage.

Benjamini-Hochberg procedure ('BH')

doi: 10.1111/j.2517-6161.1995.tb02031.x

Benjamini-Yekutieli procedure ('BY')

https://www.jstor.org/stable/2674075

Structure-Adaptive Benjamini–Hochberg Algorithm ('SABHA')

doi: 10.1111/rssb.12298

Locally Adaptive Weighting and Screening ('LAWS')

doi: 10.1080/01621459.2020.1859379

LAWS(
  pv,
  bandwidth,
  dimension = c("one", "two", "three"),
  alpha = 0.05,
  initial_filter = 0.9
)

SABHA(
  pv,
  bandwidth,
  dimension = c("one", "two", "three"),
  alpha = 0.05,
  initial_filter = 0.9
)

BH(pvals, alpha = 0.05, filter = 1)

BY(pvals, alpha = 0.05, filter = 1)

Arguments

pv, pvals
p-values

bandwidth
kernel smoothing bandwidth

dimension
the spatial dimension of underlying data. Current implementation only supports 1-3 dimensions

alpha
'FDR' level

filter, initial_filter
initial p-value filters, the goal is to remove large p-values when true signals are sparse. For LAWS and SABHA, initial_filter helps determine sparsity levels.

Value

A list of rejection results:

rejs

integer indices of rejected hypotheses;

nrejs

total number of rejections;

method

characters of method name;

filter,initial_filter

passed from arguments, used to calculate "purity" values;

pis_hat

estimated sparsity level (LAWS and SABHA only);

tau

p-value cutoff value (BH and BY only);

order,qvals

other variable from BH and BY;

bandwidth,dimension,details

other variable from LAWS and SABHA.

References

[1] Benjamini, Y. and Hochberg, Y., 1995. Controlling the false discovery rate: a practical and powerful approach to multiple testing. Journal of the Royal statistical society: series B (Methodological), 57(1), pp.289-300.

[2] Benjamini, Y. and Yekutieli, D., 2001. The control of the false discovery rate in multiple testing under dependency. Annals of statistics, pp.1165-1188.

[3] Li, A. and Barber, R.F., 2019. Multiple testing with the structure‐adaptive Benjamini–Hochberg algorithm. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 81(1), pp.45-74.

[4] Cai, T.T., Sun, W. and Xia, Y., 2020. LAWS: A Locally Adaptive Weighting and Screening Approach To Spatial Multiple Testing. Journal of the American Statistical Association, pp.1-30.